requirements.txt
# Data Versioning System Requirements
# Install with: pip install -r requirements.txt

# Core dependencies
PyYAML>=6.0          # YAML configuration file parsing
pandas>=1.3.0        # Data analysis and CSV handling

# Optional dependencies for enhanced functionality
# Uncomment as needed:
# openpyxl>=3.0.0    # Excel file support
# sqlalchemy>=1.4.0  # Database support
# psycopg2>=2.9.0    # PostgreSQL support
